# api-spec/merged/openapi.yaml
openapi: 3.0.3
info:
  title: Users Service API
  description: API for user provisioning and profile management
  version: 1.0.0
servers:
  - url: 'http://localhost:8001/api'
    description: Development server
security:
  - firebaseAuth: []
paths:
...
  /newsletter:
    get:
      summary: Get personalized newsletter
      description: |
        Returns a personalized newsletter for the authenticated user.
        Based on trips from travellers the user has interacted with.
        Requires authentication.
      operationId: getNewsletter
      tags:
        - newsletter
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Personalized newsletter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsletterResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    NewsletterTrip:
      type: object
      required:
        - tripId
        - title
        - creatorId
        - creatorName
        - relevanceReason
        - createdAt
      properties:
        tripId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        destination:
          type: string
        coverImageUrl:
          type: string
        creatorId:
          type: string
        creatorName:
          type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        relevanceReason:
          type: string
          description: Why this trip is relevant for the user
          enum:
            - creator_you_follow
            - liked_by_similar_users
            - trending_in_network
        createdAt:
          type: string
          format: date-time
    NewsletterSection:
      type: object
      required:
        - title
        - trips
      properties:
        title:
          type: string
        description:
          type: string
        trips:
          type: array
          items:
            $ref: '#/components/schemas/NewsletterTrip'
    NewsletterResponse:
      type: object
      required:
        - sections
        - generatedAt
      properties:
        sections:
          type: array
          items:
            $ref: '#/components/schemas/NewsletterSection'
        generatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
  securitySchemes:
    firebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT